requirements.txt
# requirements.txt
# 建議指定版本以確保環境一致性
# 請根據你實際安裝的版本修改版本號
Flask==2.3.2
psycopg2-binary==2.9.9